Refactored tests into abstract classes to reduce duplicate tests between classes
